requirements.txt
streamlit==0.84.0
pandas==1.2.3
numpy==1.19.5
opencv-python-headless
imageio
tensorflow-cpu==2.5.0
scikit-learn==0.23.2
spotipy==2.18.0
mtcnn==0.1.1
pickle-mixin==1.0.2
